rm unused modules and refactor
