--- addresses_to_check.txt
@@ -1,10 +1,0 @@
-# from skypeLANhome.pcap
-#10:08:b1:21:3f:cb
-44:65:0d:56:cc:d3
-#d4:dc:cd:8e:29:07
-#00:00:00:00:00:00
-#172.17.0.3
-#UDP 192.168.0.102:18826 192.168.0.105:26454 
-#TCP 192.168.6.105:26454 192.168.0.102:2
-#192.168.0.105
-#UDP 192.168.122.20:4000 192.168.122.10:16002
